extract FeatureOverlay
