+++ garibaldi/spikes/overlay-biscuits/src/MapView.js
@@ -3,54 +3,8 @@ import { useContext } from "react";
 import { MapBoxContext } from "./MapBoxContext";
 import { useRef, useLayoutEffect, useState } from "react";
 import ReactMapGL from "react-map-gl";
-import { CanvasOverlay } from "react-map-gl";
-import { LngLatBounds, LngLat } from "mapbox-gl";
-import { geoPath, geoTransform } from "d3-geo";
-
-function FeatureOverlay({ boundingBox, featureLoader }) {
-  function redraw({ width, height, ctx, isDragging, project, unproject }) {
-    const center = project(boundingBox.getCenter().toArray());
-    ctx.clearRect(0, 0, width, height);
-    ctx.beginPath();
-    ctx.arc(center[0], center[1], 10.0, 0, 2 * Math.PI, false);
-    ctx.fillStyle = "green";
-    ctx.fill();
-
-    const geoJsonBounds = geoJsonBoundsFromLngLatBounds(boundingBox);
-
-    const reticuleProjection = geoTransform({
-      point: function(lon, lat) {
-        const point = project(new LngLat(lon, lat).toArray());
-        this.stream.point(point[0], point[1]);
-      }
-    });
-
-    const generator = geoPath()
-      .projection(reticuleProjection)
-      .context(ctx);
-
-    if (!isDragging) {
-      console.time("redraw: all");
-      const features = featureLoader();
-      const geoJson = { type: "FeatureCollection", features };
-
-      ctx.lineWidth = 1;
-      ctx.beginPath();
-      ctx.strokeStyle = "blue";
-      generator(geoJson);
-      ctx.stroke();
-      console.timeEnd("redraw: all");
-    }
-
-    ctx.lineWidth = 2;
-    ctx.beginPath();
-    ctx.strokeStyle = "red";
-    generator(geoJsonBounds);
-    ctx.stroke();
-  }
-
-  return <CanvasOverlay redraw={redraw} />;
-}
+import { LngLatBounds } from "mapbox-gl";
+import { FeatureOverlay } from "./FeatureOverlay";
 
 function reticuleFromMapBounds(bounds) {
   const northSouthExtent = bounds.getSouth() - bounds.getNorth();
@@ -70,29 +24,6 @@ function reticuleFromMapBounds(bounds) {
   ]);
 
   return reticuleBounds;
-}
-
-function geoJsonBoundsFromLngLatBounds(bounds) {
-  return {
-    type: "FeatureCollection",
-    features: [
-      {
-        type: "Feature",
-        geometry: {
-          type: "Polygon",
-          coordinates: [
-            [
-              bounds.getNorthEast().toArray(),
-              bounds.getSouthEast().toArray(),
-              bounds.getSouthWest().toArray(),
-              bounds.getNorthWest().toArray(),
-              bounds.getNorthEast().toArray()
-            ]
-          ]
-        }
-      }
-    ]
-  };
 }
 
 export function MapView({ city }) {
